pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "udva-backend"
version = "0.1.0"
description = "Udva — AI search visibility platform backend"
requires-python = ">=3.12"
dependencies = [
    # Web framework
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",

    # Validation & settings
    "pydantic[email]>=2.9.0",
    "pydantic-settings>=2.6.0",

    # Database
    "sqlalchemy[asyncio]>=2.0.36",
    "asyncpg>=0.30.0",
    "alembic>=1.14.0",

    # Task queue
    "celery[redis]>=5.4.0",
    "redis>=5.2.0",

    # Auth
    "fastapi-users[sqlalchemy]>=13.0.0",
    "python-jose[cryptography]>=3.3.0",
    "bcrypt>=4.2.0",

    # Payments & email
    "dodopayments>=1.4.0",
    "standardwebhooks>=1.0.0",
    "resend>=2.4.0",

    # Error tracking
    "sentry-sdk>=2.19.0",

    # LLM clients
    "openai>=1.54.0",
    "anthropic>=0.39.0",
    "google-generativeai>=0.8.3",

    # HTTP & data sources
    "httpx>=0.28.0",
    "praw>=7.8.1",
]

[tool.uv]
dev-dependencies = [
    "ruff>=0.8.0",
    "mypy>=1.13.0",
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
    "aiosqlite>=0.20.0",
    "types-python-jose>=3.3.4",
    "types-redis>=4.6.0",
]

[tool.ruff]
target-version = "py312"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "SIM"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.12"
strict = true
ignore_missing_imports = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
